updated styling

+++ src/components/Comment/Comment.js
@@ -8,38 +8,56 @@ import './Comment.scss';
 class Comment extends Component {
 
   state = {
-    relDate: LastSeen(this.props.timestamp)
+    relDate: LastSeen(this.props.timestamp),
+    isDeleting: false,
   };
 
   // click handler for deleting comments
   handleDelete = () => {
-    const errorMessage = < p > Error fetching data, please try reloading in a few moments</p >;
+    if (this.state.isDeleting) {
+      return;
+    }
+
+    const shouldDelete = window.confirm("Delete this comment?");
+    if (!shouldDelete) {
+      return;
+    }
 
     const { videoId, commentId, updateActiveVideoObj } = this.props;
+    this.setState({ isDeleting: true });
+
     // API DELETE req sending vidID and commentID then updates active vid Obj
     apiUtils.deleteVideoComment(videoId, commentId)
       .then(res => updateActiveVideoObj(res.data))
       .catch(err => {
         console.log(err);
-        return errorMessage;
+        this.setState({ isDeleting: false });
       });
   };
 
   render() {
 
-    const { relDate } = this.state;
+    const { relDate, isDeleting } = this.state;
     const { name, comment } = this.props;
 
     return (
       <div className='comment'>
         <div className='comment__left'>
           <div className="comment__avatar--blank"></div>
-          <img
+          <button
+            type="button"
             onClick={() => this.handleDelete()}
-            src={deleteIcon}
-            alt="delete icon"
-            className='comment__icon'
-          />
+            className='comment__icon-button'
+            aria-label={`Delete comment from ${name}`}
+            disabled={isDeleting}
+          >
+            <img
+              src={deleteIcon}
+              alt=""
+              aria-hidden="true"
+              className='comment__icon'
+            />
+          </button>
         </div>
         <div className="comment__right">
           <h3 className="comment__name">{name}</h3>
